+++ packages/devtools_app/lib/src/service/service_manager.dart
@@ -18,6 +18,7 @@ import '../shared/diagnostics/inspector_service.dart';
 import '../shared/error_badge_manager.dart';
 import '../shared/feature_flags.dart';
 import '../shared/globals.dart';
+import '../shared/server/server.dart' as server;
 import '../shared/title.dart';
 import '../shared/utils.dart';
 import 'service_registrations.dart' as registrations;
@@ -130,6 +131,13 @@ class ServiceConnectionManager {
     if (debugLogServiceProtocolEvents) {
       serviceTrafficLogger = VmServiceTrafficLogger(service!);
     }
+
+    unawaited(
+      server.notifyForVmServiceConnection(
+        vmServiceUri: serviceManager.serviceUri!,
+        connected: true,
+      ),
+    );
   }
 
   void _beforeCloseVmService(VmServiceWrapper? service) {
@@ -140,6 +148,15 @@ class ServiceConnectionManager {
         ? OfflineConnectedApp.parse(serviceManager.connectedApp!.toJson())
         : null;
     offlineController.previousConnectedApp = previousConnectedApp;
+
+    // This must be called before we close the VM service so that
+    // [serviceManager.serviceUri] is not null.
+    unawaited(
+      server.notifyForVmServiceConnection(
+        vmServiceUri: serviceManager.serviceUri!,
+        connected: false,
+      ),
+    );
   }
 
   void _afterCloseVmService(VmServiceWrapper? service) {